KeyRange::prefix: handle empty values gracefully
To request all keys from Etcd a special KeyRange value containing null
bytes is prepared by `KeyRange::all`. When `KeyRange::prefix` is called
with an empty string, the expected behaviour is for it to behave as
`KeyRange::all` does. This commit makes `KeyRange::prefix` return the
special 'no-prefix' `KeyRange` when the input is empty.